pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "isek"
version = "0.1.0"
description = "ISEK Distributed Multi-Agent Framework"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "ISEK Team", email = "sparks@isek.xyz" }
]
requires-python = ">=3.8"
keywords = ["LLM", "multi-agent", "distributed", "AI"]

dependencies = [
    "click>=8.0.0",
    "loguru>=0.6.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    # Core dependencies
    "protobuf>=3.20.0,<4.0.0",
    "grpcio>=1.46.0",            
    "grpcio-tools>=1.46.0",
    "openai>=0.27.0",
    "flask>=2.0.0",
    "etcd3>=0.12.0"
]

[project.scripts]
isek = "isek.cli:cli"